Added adding function

+++ src/main/java/ua/alexd/CarRentService/service/RentService.java
@@ -67,8 +67,7 @@ public class RentService {
 
     public boolean deleteRent(String id) {
         var deletionRent = getRentById(id);
-        if (deletionRent.isPresent()) {
-            rentTriggers.releaseCar(deletionRent.get());
+        if (deletionRent.isPresent() && rentTriggers.releaseCar(deletionRent.get())) {
             rentRepository.delete(deletionRent.get());
             return true;
         }
